Adds more thorough Facebook integration
Only users who are authorized to use the app can log in.  These users are hardcoded into the database.
Trying to navigate to the backlog/:name url when not logged in to Facebook under an authorized account will result in a redirect to the sign in page
Logging in with an unauthorized account results in a navigation to the User Unauthorized page.